+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # Changelog
 
+## 2.3.0 --- Multiplexing Catarsis (WIP)
+
+- Allow multiple different optimization targets to be reported
+- Fix selected community not being reported
+
 ## 2.2.0 --- Consistency Crysis (Feb 21, 2023)
 
 - Use a separate environment for the MILP and the LPs